LayoutTests/editing/undomanager/undomanager-use-after-free-expected.txt

Ensure that undoManager doesn't have use-after-free vulnerabilities.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


call undoManager.transact() during the execute callback
PASS undoManager.transact(transaction) threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
call undoManager.undo() during the undo callback
PASS undoManager.undo() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
call undoManager.clearRedo() during the redo callback
PASS undoManager.redo() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
call undoManager.transact() with a transaction that disconnects the undoManager when undoing
PASS result is "execute "
call undoManager.undo()
PASS result is "execute undo "
PASS edit.undoManager is null
PASS undoManager.length is 0
call undoManager.redo() when the undoManager is disconnected
PASS undoManager.redo() threw exception Error: INVALID_ACCESS_ERR: DOM Exception 15.
PASS result is "execute undo "
PASS successfullyParsed is true

TEST COMPLETE
